requirements.txt
textblob
nltk
newspaper3k
streamlit
tk